requirements.txt
kafka-python==2.1.5
tzlocal==4.3.1